FIX: base of iterator was deleted, now we are free!

+++ lab_02/lab_02/main.cpp
@@ -3,7 +3,7 @@
 
 int main(void)
 {
-    Vector<int> initializerListExample = {1, 2, 3, 4, 5};
+    Vector<double> initializerListExample = {1, 2, 3, 4, 5};
     Vector<float> lenExample(3);
     Vector<double> lenFillExample(4, 6.66);
     Vector<long double> emptyVector(0);
@@ -12,4 +12,5 @@ int main(void)
     std::cout<< "lenInit: "<< lenExample<< std::endl;
     std::cout<< "LenFillInit with 6.66 filler: "<< lenFillExample<< std::endl;
     std::cout<< "Empty vector: "<< emptyVector<< std::endl;
+
 }
